+++ src/PSRule/Pipeline/GetRulePipeline.cs
@@ -7,15 +7,15 @@ namespace PSRule.Pipeline
 {
     public sealed class GetRulePipeline : RulePipeline
     {
-        internal GetRulePipeline(PSRuleOption option, string[] path, RuleFilter filter, PipelineContext context)
-            : base(context, option, path, filter)
+        internal GetRulePipeline(PSRuleOption option, RuleSource[] source, RuleFilter filter, PipelineContext context)
+            : base(context, option, source, filter)
         {
             // Do nothing
         }
 
         public IEnumerable<Rule> Process()
         {
-            return HostHelper.GetRule(_Option, _Path, _Filter);
+            return HostHelper.GetRule(_Option, _Source, _Filter);
         }
     }
 }