+++ tests/PSRule.Tests/TestModule/rules/Test.Rule.ps1
@@ -1,0 +1,7 @@
+#
+# A set of test rules in a module
+#
+
+Rule 'Rule1' {
+    # This is a test rule
+}